fix: extract related pr in merge action (#35)
Merge Action won't always be triggered on a PR event. But if we can, we
should still try and merge the PR if the conditions match. This means
extracting the related PR for the event.

This PR adds a `related_pr` method to GithubEvent which tries to dig out
the PR info from the different event types.

+++ github-api/src/models/pull_request.rs
@@ -71,7 +71,7 @@ pub struct PullRequest {
     pub review_comment_url: Url,
     pub comments_url: Url,
     pub statuses_url: Url,
-    pub number: usize,
+    pub number: u64,
     pub state: State,
     pub locked: bool,
     pub title: String,
@@ -128,8 +128,8 @@ pub struct IssuePullRequest {
 pub struct CheckRunPullRequest {
     pub base: GitReferenceShort,
     pub head: GitReferenceShort,
-    pub id: usize,
-    pub number: usize,
+    pub id: u64,
+    pub number: u64,
     pub url: Url,
 }
